FOIA-87: Moved import setup functions to Agency Report XML Upload form.
